fix: 매니지먼트 모듈 리팩토링중
- 로그인 횟수 패스포트에 포함
- 전체적인 로직 수정 및 몽 히스토리 로직 추가 (AOP 로 코드 수정이 필요)

+++ core/src/main/java/com/mongs/core/vo/mqtt/PublishFeedVo.java
@@ -10,6 +10,6 @@ public record PublishFeedVo(
         Double satiety,
         Double health,
         Double sleep,
-        Integer exp
+        Double exp
 ) {
 }
