+++ management/src/main/java/com/mongs/management/domain/mong/controller/dto/response/FeedMongResDto.java
@@ -10,6 +10,6 @@ public record FeedMongResDto(
         Double satiety,
         Double strength,
         Double sleep,
-        Integer exp
+        Double exp
 ) {
 }
