+++ management/src/main/java/com/mongs/management/domain/mong/controller/dto/response/PoopCleanResDto.java
@@ -6,6 +6,6 @@ import lombok.Builder;
 public record PoopCleanResDto(
         Long mongId,
         Integer poopCount,
-        Integer exp
+        Double exp
 ) {
 }
